Dissolve PFBTestCase and TestPFB
